shrink int to auto-map to int8 (#523)

+++ +file/mapType.m
@@ -43,7 +43,7 @@ switch dtype
     case 'long'
         dt = 'int64';
     case 'int'
-        dt = 'int32';
+        dt = 'int8';
     case 'short'
         dt = 'int16';
     otherwise
